feature: Largest Contentful Paint (LCP)
Ensuring that the server side rendered page and rehydrated page are
identical is extremely important to page speed. The change to the
`<SearchBar />` brings the LCP inline with the FP. This cuts the LCP
time down from 50%-70% depending on device.

Additionally, the previously mentioned changes alogn with removing the
`<BreakpointVisualizer />` resovldes the react warnings:

   "Unexpected div in a div"